pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lachasseauxtresor"
version = "1.0.0"
description = "CrewAI-based adventure book generator inspired by La Chasse au Trésor TV show"
authors = [{name = "La Chasse au Trésor Team"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "crewai>=0.30.0",
    "crewai-tools>=0.4.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "jsonschema>=4.20.0",
    "click>=8.1.7",
    "rich>=13.7.0",
    "tqdm>=4.66.0",
    "loguru>=0.7.2"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "flake8>=6.1.0",
    "mypy>=1.8.0"
]

[project.scripts]
lachasseauxtresor = "src.main:cli"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]